+++ CHANGELOG.md
@@ -5,10 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.5] - 2021-02-26
+* Update MultiLabel to use js private encapsulation rather than constructor encapsulation.
+* MultiLabel and ImageLabel can be created with or without the `new` keyword.
+
 ## [2.0.0] - 2021-02-22
 * Images will be generated using the dom in place of using cc.Label in order to work around poor drivers on certain 
 low-cost Lenovo Chromebooks.
-* createLabel was replaced with ImageLabel()
+* createLabel was replaced with ImageLabel().
 
 ## [Unreleased]